+++ src/main.py
@@ -3,9 +3,12 @@ import os
 import sys
 
 from dotenv import load_dotenv
+from fastapi import FastAPI
+
+import texto
 
 sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
-from utils import Constantes, setup_logging
+from utils import setup_logging
 
 # Configura el logging
 setup_logging()
@@ -15,12 +18,5 @@ logger.setLevel("INFO")
 # Cargar .env
 load_dotenv()
 
-if __name__ == "__main__":
-    print(Constantes.encoding.value)
-    print(os.getenv("token_hf"))
-    logger.info("Prueba info")
-    logger.error("Prueba error")
-    try:
-        donuts_per_guest = 5 / 0
-    except ZeroDivisionError:
-        logger.exception("Prueba excepción")
+app = FastAPI()
+app.include_router(texto.router)